process_key_to_parser_function created
