+++ src/parsers/pilot_talent.py
@@ -3,14 +3,13 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from parsers.object import Object, ParseAction, ParseTarget
-
+from parsers.object import Object
 from parsers.localization_table import parse_localization
 from parsers.image import parse_image_asset_path
 from parsers.module_stat import ModuleStat
 from parsers.module_tag import ModuleTag
 
-from utils import asset_path_to_data, asset_path_to_file_path, get_json_data, parse_colon_colon
+from utils import ParseTarget, ParseAction, asset_path_to_data, asset_path_to_file_path, get_json_data, parse_colon_colon
 
 class PilotTalent(Object):
     objects = dict()  # Dictionary to hold all PilotTalent instances
